I add a comment to illustrate use of functions
The cachematrix.R file contains two functions, makeCacheMatrix() and cacheSolve().
The first function in the file, makeCacheMatrix() creates an R object that stores a matrix and its inverse.
The second function, cacheSolve() requires an argument that is returned by makeCacheMatrix()
in order to retrieve the inverse from the cached value that is stored in the makeCacheMatrix() object's environment.